refactor

+++ backend/controllers/product.controller.js
@@ -74,6 +74,13 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res.status(404).json({
+      success: false,
+      message: "Product not found",
+    });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     return res.status(200).json({
@@ -82,9 +89,9 @@ export const deleteProduct = async (req, res) => {
     });
   } catch (error) {
     console.error("Error deleting product", error);
-    return res.status(404).json({
+    return res.status(500).json({
       success: false,
-      message: "Product not found",
+      message: "Server error",
     });
   }
 };
